+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Bump `actions/setup-python` to 5.4.0
 - Update minor and patch NPM dependencies
+- Improve code coverage on core and modules components
 
 ### Fixed